backend/requirements.txt
# FastAPI and ASGI
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
python-multipart>=0.0.9

# Database
sqlalchemy>=2.0.36
asyncpg>=0.30.0
alembic>=1.14.0
aiosqlite>=0.20.0

# Redis
redis>=5.2.0

# MQTT
paho-mqtt>=2.1.0

# SSH Tunnel
paramiko>=3.5.0
sshtunnel>=0.4.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.2.1

# Validation
pydantic>=2.10.0
pydantic-settings>=2.6.0
email-validator>=2.2.0

# Utilities
python-dotenv>=1.0.1
pytz>=2024.2
httpx>=0.28.0
loguru>=0.7.3

# Development
pytest>=8.3.0
pytest-asyncio>=0.24.0
pytest-cov>=6.0.0
black>=24.10.0
isort>=5.13.2
flake8>=7.1.0
